test: split surface continuity tests
